refactor: add-card 페이지 폴더 명 변경
