+++ src/pages/add-card/components/AddCardPreview/utils/index.ts
